pyproject.toml
[project]
name = "opendart"
version = "0.1.0"
description = "ETL system for Korean financial data from Open DART API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opendartreader>=0.2.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "python-dotenv>=1.0",
    "click>=8.0",
    "pandas>=2.0",
    "apscheduler>=3.10",
]

[project.scripts]
opendart = "opendart.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opendart"]

[tool.uv]
package = true
